# assets/openapi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "https://staging.spacebar.chat/api/",
            "description": "Official Spacebar Instance"
        }
    ],
    "info": {
        "description": "Spacebar is a free open source selfhostable discord compatible chat, voice and video platform",
        "version": "1.0.0",
        "title": "Spacebar HTTP API Routes",
        "termsOfService": "",
        "contact": {
            "name": "Spacebar"
        },
        "license": {
            "name": "AGPLV3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Bearer/Bot prefixes are not required."
            }
        },
        "schemas": {
...
            "Embed": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "enum": [
                            "article",
                            "gifv",
                            "image",
                            "link",
                            "rich",
                            "video"
                        ],
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "color": {
                        "type": "integer"
                    },
                    "footer": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "icon_url": {
                                "type": "string"
                            },
                            "proxy_icon_url": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "text"
                        ]
                    },
                    "image": {
                        "$ref": "#/components/schemas/EmbedImage"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/EmbedImage"
                    },
                    "video": {
                        "$ref": "#/components/schemas/EmbedImage"
                    },
                    "provider": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            },
                            "icon_url": {
                                "type": "string"
                            },
                            "proxy_icon_url": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "value": {
                                    "type": "string"
                                },
                                "inline": {
                                    "type": "boolean"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "name",
                                "value"
                            ]
                        }
                    }
                }
            },
            "EmbedImage": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "proxy_url": {
                        "type": "string"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "width": {
                        "type": "integer"
                    }
                }
            },
...
            "MessageCreateSchema": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "integer"
                    },
                    "content": {
                        "type": "string"
                    },
                    "nonce": {
                        "type": "string"
                    },
                    "channel_id": {
                        "type": "string"
                    },
                    "tts": {
                        "type": "boolean"
                    },
                    "flags": {
                        "type": "string"
                    },
                    "embeds": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Embed"
                        }
                    },
                    "embed": {
                        "$ref": "#/components/schemas/Embed"
                    },
                    "allowed_mentions": {
                        "type": "object",
                        "properties": {
                            "parse": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "roles": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "users": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "replied_user": {
                                "type": "boolean"
                            }
                        },
                        "additionalProperties": false
                    },
                    "message_reference": {
                        "type": "object",
                        "properties": {
                            "message_id": {
                                "type": "string"
                            },
                            "channel_id": {
                                "type": "string"
                            },
                            "guild_id": {
                                "type": "string"
                            },
                            "fail_if_not_exists": {
                                "type": "boolean"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "channel_id",
                            "message_id"
                        ]
                    },
                    "payload_json": {
                        "type": "string"
                    },
                    "file": {
                        "type": "object",
                        "properties": {
                            "filename": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "filename"
                        ]
                    },
                    "attachments": {
                        "description": "TODO: we should create an interface for attachments\nTODO: OpenWAAO<-->attachment-style metadata conversion",
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "filename": {
                                    "type": "string"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "filename",
                                "id"
                            ]
                        }
                    },
                    "sticker_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "BulkDeleteSchema": {
                "type": "object",
                "properties": {
                    "messages": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "messages"
                ]
            },
...
            "MessageEditSchema": {
                "type": "object",
                "properties": {
                    "file": {
                        "type": "object",
                        "properties": {
                            "filename": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "filename"
                        ]
                    },
                    "embed": {
                        "$ref": "#/components/schemas/Embed"
                    },
                    "flags": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "nonce": {
                        "type": "string"
                    },
                    "channel_id": {
                        "type": "string"
                    },
                    "tts": {
                        "type": "boolean"
                    },
                    "embeds": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Embed"
                        }
                    },
                    "allowed_mentions": {
                        "type": "object",
                        "properties": {
                            "parse": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "roles": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "users": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "replied_user": {
                                "type": "boolean"
                            }
                        },
                        "additionalProperties": false
                    },
                    "message_reference": {
                        "type": "object",
                        "properties": {
                            "message_id": {
                                "type": "string"
                            },
                            "channel_id": {
                                "type": "string"
                            },
                            "guild_id": {
                                "type": "string"
                            },
                            "fail_if_not_exists": {
                                "type": "boolean"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "channel_id",
                            "message_id"
                        ]
                    },
                    "payload_json": {
                        "type": "string"
                    },
                    "attachments": {
                        "description": "TODO: we should create an interface for attachments\nTODO: OpenWAAO<-->attachment-style metadata conversion",
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "filename": {
                                    "type": "string"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "filename",
                                "id"
                            ]
                        }
                    },
                    "sticker_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    },
...
    "paths": {
...
        "/users/@me/delete/": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
...
        "/guilds/{guild_id}/delete/": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/channels/{channel_id}/messages/bulk-delete/": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkDeleteSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
...
        "/channels/{channel_id}/messages/{message_id}/": {
            "patch": {
                "x-right-required": "SEND_MESSAGES",
                "x-permission-required": "SEND_MESSAGES",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageEditSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "put": {
                "x-right-required": "SEND_BACKDATED_EVENTS",
                "x-permission-required": "SEND_MESSAGES",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageCreateSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "get": {
                "x-permission-required": "VIEW_CHANNEL",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
...
        "/applications/{id}/delete": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "id"
                    }
                ],
                "tags": [
                    "applications"
                ]
            }
        },
...
        "/users/{id}/delete/": {
            "post": {
                "x-right-required": "MANAGE_USERS",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "id"
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
...
    }
}